now the values are only size_t

+++ interpreter.h
@@ -6,5 +6,5 @@
 void Interpret(InterpreterWindow& window);
 
 void PushElem(PostfixElem el);
-int PopVal();
-void SetVarAndPop(int val);
+size_t PopVal();
+void SetVarAndPop(size_t val);
